# server/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.BulkIDsRequest:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
    type: object
...
  models.InviteToken:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      email:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      token:
        type: string
      usedAt:
        type: string
      usedBy:
        type: string
    type: object
...
  models.OverviewResponse:
    properties:
      activityTrend:
        items:
          $ref: '#/definitions/models.DayActivity'
        type: array
      health:
        $ref: '#/definitions/models.OverviewHealth'
      instanceInfo:
        $ref: '#/definitions/models.InstanceInfo'
      kpis:
        $ref: '#/definitions/models.OverviewKPIs'
      needsAttention:
        items:
          $ref: '#/definitions/models.AttentionItem'
        type: array
      recentRoomEvents:
        items:
          $ref: '#/definitions/models.RoomEvent'
        type: array
      recentSignups:
        items:
          $ref: '#/definitions/models.RecentUser'
        type: array
      roomComposition:
        $ref: '#/definitions/models.RoomComposition'
    type: object
...
host: localhost:8090
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a Bedrud Backend API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Bedrud Backend API
  version: "1.0"
paths:
  /admin/invite-tokens:
    get:
      consumes:
      - application/json
      description: Get paginated list of registration invite tokens. Superadmin access
        required.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{tokens, total}'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to fetch tokens
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List invite tokens
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Create a new registration invite token. Superadmin access required.
      parameters:
      - description: '{email?: string, expiresInHours?: int}'
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created token
          schema:
            $ref: '#/definitions/models.InviteToken'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to create token
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create invite token
      tags:
      - admin
  /admin/invite-tokens/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a registration invite token. Superadmin access required.
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{status: success}'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to delete token
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete invite token
      tags:
      - admin
...
  /admin/online-count:
    get:
      consumes:
      - application/json
      description: Get total online participants and publishers across all active
        LiveKit rooms. Superadmin access required.
      produces:
      - application/json
      responses:
        "200":
          description: '{totalParticipants, totalPublishers, activeRooms, rooms}'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get online count
      tags:
      - admin
  /admin/overview:
    get:
      consumes:
      - application/json
      description: |-
        Aggregated system stats for the admin dashboard: health status, KPI metrics,
        7-day activity trend, room composition breakdown, and recent user/room activity.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OverviewResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin dashboard overview
      tags:
      - admin
  /admin/queue:
    get:
      consumes:
      - application/json
      description: |-
        Real-time queue status including pending/active/done/failed counts,
        processed rates, and recent failure diagnostics.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QueueStats'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Queue statistics
      tags:
      - admin
  /admin/recordings:
    get:
      consumes:
      - application/json
      description: List all recordings with optional filters. Superadmin access required.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 1000
        description: Items per page
        in: query
        maximum: 1000
        name: perPage
        type: integer
      - description: Filter by room ID
        in: query
        maxLength: 36
        name: roomId
        type: string
      - description: Filter by status (completed, processing, failed, deleting)
        in: query
        name: status
        type: string
      - description: Filter by ISO 8601 timestamp
        in: query
        name: createdAfter
        type: string
      - description: Filter by ISO 8601 timestamp
        in: query
        name: createdBefore
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{recordings, total, page, limit}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid filter parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to list recordings
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Admin list recordings
      tags:
      - admin
  /admin/recordings/bulk/delete:
    post:
      consumes:
      - application/json
      description: Mark recordings as deleting and enqueue async deletion jobs. Superadmin
        access required. Max 500 IDs per request.
      parameters:
      - description: Recording IDs to delete
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.BulkIDsRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Deletion queued
          schema:
            $ref: '#/definitions/handlers.BulkResult'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to process
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Bulk delete recordings
      tags:
      - admin
...
  /admin/rooms/{roomId}/participants:
    get:
      consumes:
      - application/json
      description: Get live participant data from LiveKit for a specific room. Superadmin
        access required.
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{participants, room}'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get room participants
      tags:
      - admin
...
  /admin/rooms/{roomId}/reactivate:
    post:
      consumes:
      - application/json
      description: Reactivate a suspended room. Superadmin access required.
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Reactivated room
          schema:
            $ref: '#/definitions/models.Room'
        "400":
          description: Room not suspended
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to reactivate
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Reactivate room
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"